+++ NEWS.md
@@ -2,6 +2,7 @@
 ----------------
 
 - New defcustom, `org-d20-letter-monsters`
+- New defcustom, `org-d20-continue-monster-numbering`
 
 0.2 (2018-12-31)
 ----------------